This PR has a bunch of performance improvements, most notably it only creates the regex once. Overall this increases the performance for me by ~30%.

master
```
lodash.get x 139,273 ops/sec ±14.20% (64 runs sampled)
fast-get x 359,675 ops/sec ±26.46% (72 runs sampled)
```

PR
```
lodash.get path x 752,851 ops/sec ±5.86% (68 runs sampled)
lodash.get string x 146,373 ops/sec ±27.13% (70 runs sampled)
fast-get path x 617,748 ops/sec ±23.65% (67 runs sampled)
fast-get string x 466,544 ops/sec ±1.88% (83 runs sampled)
```